implement ServiceType CRUD api

+++ src/main/java/com/sliit/autocarepro/Model/ServiceType.java
@@ -13,38 +13,38 @@ public class ServiceType {
         this.description = description;
         this.timeRequired = timeRequired;
         this.priceRange = priceRange;
-    }
+        }
 
-    public int getId() {
-        return id;
-    }
-    public void setId(int id) {
-        this.id = id;
-    }
+        public long getId() {
+            return id;
+        }
+        public void setId(int id) {
+            this.id = id;
+        }
 
-    public String getService() {
-        return service;
-    }
-    public void setService(String service) {
-        this.service = service;
-    }
-    public String getDescription() {
-        return description;
-    }
-    public void setDescription(String description) {
-        this.description = description;
-    }
-    public String getTimeRequired() {
-        return timeRequired;
-    }
-    public void setTimeRequired(String timeRequired) {
-        this.timeRequired = timeRequired;
-    }
-    public String getPriceRange() {
-        return priceRange;
-    }
-    public void setPriceRange(String priceRange) {
-        this.priceRange = priceRange;
-    }
+        public String getService() {
+            return service;
+        }
+        public void setService(String service) {
+            this.service = service;
+        }
+        public String getDescription() {
+            return description;
+        }
+        public void setDescription(String description) {
+            this.description = description;
+        }
+        public String getTimeRequired() {
+            return timeRequired;
+        }
+        public void setTimeRequired(String timeRequired) {
+            this.timeRequired = timeRequired;
+        }
+        public String getPriceRange() {
+            return priceRange;
+        }
+        public void setPriceRange(String priceRange) {
+            this.priceRange = priceRange;
+        }
 
-}
+    }
